requirements.txt
aiogram==2.10.1
aiohttp==3.7.3
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptg==0.2.post1
ffmpeg==1.4
hachoir==3.1.1
idna==2.8
imutils==0.5.3
instaloader==4.5.3
lxml==4.5.2
multidict==4.7.6
numpy==1.18.5
opencv-python==4.2.0.34
prettytable==0.7.2
protobuf==3.12.2
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
Pyrogram~=1.0.7
PySocks==1.7.1
pyTelegramBotAPI==3.7.1
pytube==9.6.0
pytube3==9.6.4
pytz==2020.1
requests==2.24.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
Telethon==1.14.0
TgCrypto==1.2.1
typing-extensions==3.7.4.2
uhttplib==0.2
urllib3==1.25.9
yarl==1.4.2
youtube_dl==2021.4.7
aiofiles==0.6.0
uWSGI==2.0.19.1
lxml==4.5.2
python-dotenv~=0.14.0
proxybroker==0.3.2
asyncio==3.4.3
free-proxy==1.0.2
aiohttp-socks==0.5.5
mysql-connector-python==8.0.22
pafy==0.5.5